Moved profile components to folder components in profile
