+++ IzvorniKod/frontend/src/pages/Profile/components/StatisticsComponent/StatisticsComponent.tsx
